examples/images/sensor/requirements.txt
Flask
requests
iperf3
paho-mqtt
psutil
